+++ architecture_tool_django/utils/confluence_wrapper.py
@@ -7,12 +7,6 @@ import re
 from atlassian import Confluence
 from django.conf import settings
 
-private_auth = {
-    "username": settings.CONFLUENCE_USER,
-    "password": settings.CONFLUENCE_PASS,
-}
-extra_headers = {"keyid": settings.API_KEY, "X-Atlassian-Token": "nocheck"}
-
 
 def tiny_to_page_id(tiny):
     # https://community.atlassian.com/t5/Confluence-questions/What-is-the-algorithm-used-to-create-the-quot-Tiny-links-quot/qaq-p/186555
@@ -29,9 +23,16 @@ class MyConfluence(Confluence):
         url = settings.CONFLUENCE_URL
         if "url" not in kwargs:
             kwargs["url"] = url
-        kwargs.update(private_auth)
+        kwargs.update(
+            {
+                "username": settings.CONFLUENCE_USER,
+                "password": settings.CONFLUENCE_PASS,
+            }
+        )
         super().__init__(*args, **kwargs)
-        self.default_headers.update(extra_headers)
+        self.default_headers.update(
+            {"keyid": settings.API_KEY, "X-Atlassian-Token": "nocheck"}
+        )
 
     def request(self, *args, **kwargs):
         # The API is available directly under the gateway address instead of under rest/api
@@ -40,7 +41,7 @@ class MyConfluence(Confluence):
         # (method defined as keyword, but called as positional parameter)
         if len(args) > 0:
             kwargs["method"] = args[0]
-        headers = extra_headers.copy()
+        headers = {"keyid": settings.API_KEY, "X-Atlassian-Token": "nocheck"}
         if kwargs["headers"]:
             headers.update(kwargs["headers"])
         if not ("method" in kwargs and kwargs["method"] == "PUT"):